This is a first try at adding multiple AES key to viaccess reader. I added a new variable so that the previous one still works. the new variables is AESKeys (the old one is AESKey .. without the 's'). as the D2 nano reference multiple keys, starting at index 0 here is how to specify the new keys: AESKeys=CAID@IDENT:Key0,key1,key2, ..,KeyN;CAID@IDENT:Key0,key1,key2, ..,KeyN here is an exemple : AESKeys=0500@030B00:0,0,AESAESAESAESAESAESAESAESAESAESAE;0500@032820:AESAESAESAESAESAESAESAESAESAESAE,0;0500@032830:AESAESAESAESAESAESAESAESAESAESAE
so if you only have the key for tntsat SD channela :
AESKeys=0500@030B00:0,0,EF............................E3
this means that we will use EF..E3 as the key for the D2 nano for the SD channel
here is what we know about keys for now :

TntSat:
    SD D2 nano : D2 02 0D 02 -> key #2 , works with current AES key 
    HD D2 nano : D2 02 0D 00 -> key #0 , need new AES key

TNTop :
    D2 02 0D 01 -> key #1 , key is unknow as of now

CSat :
    SD no D2 nano.
    HD D2 02 0D 00 -> Key #0, use CSAT AES key -> looks like it's been changed. same key as TntSat ?

As I don't have tntsat or csat it's hard to test. So if someone could test this for me and report.



git-svn-id: http://www.streamboard.tv/svn/oscam/trunk@2722 4b0bc96b-bc66-0410-9d44-ebda105a78c1