webif: Create and use private templates array.
This allows storing of additional information along with templates
data (which is const).

For now we store name hash but in the future the data that comes from
struct templates (in webif/pages.c) may need to be preprocessed
(uncompressed for example) and the abstraction in this commit makes
it easier to do such changes.

git-svn-id: http://www.streamboard.tv/svn/oscam/trunk@8434 4b0bc96b-bc66-0410-9d44-ebda105a78c1