Fix showing active conax entitlements
On the last day when card is active (eg. on 30rd of the month, and when
the entitlement date is also 30rd) the webif is showing 'no entitlements'.
It should be to that date, including.
The commit fixes this problem and also some timezone mess.

When the tm_isdst was 0, then loaded entitlements was with DST correction
thus leading to loading wrong dates (with hour corrections depending on
entitlement date).
The current date on webif was rounded to the whole day as UTC.

Now active entitlements checks (status and entitlements pages) are all
based on local time.



git-svn-id: http://www.streamboard.tv/svn/oscam/trunk@8947 4b0bc96b-bc66-0410-9d44-ebda105a78c1