[gbx] Introduce optional parameter cccam_reshare in gbox section of oscam.server.
cccam_reshare can be used to control the number of hops of cccam cards in the gbx network.
It is independent from the global reshare settings of cccam and is similar to the parameter
cccreshare in the reader settings of cccam. 
However the logic is gbx-like not cccam.
Setting it to 0 will prevent cccam cards from being shared. Settings > 0 represent the
number of hops.



git-svn-id: http://www.streamboard.tv/svn/oscam/trunk@10367 4b0bc96b-bc66-0410-9d44-ebda105a78c1