List destruction should always set the list to NULL.
This removes ll_destroy_xxx_NULL() macros in favour of making all
ll_destroyXX() functions set the list parameter to NULL. We already
have three list destruction functions (too many) in addition to
that having 2 more (in the form of macros) was starting to get
unwise.

This change also removes the need for this idiom (used in couple
of places):

{{{
  ll_destroy(x);
  x = NULL;
}}}

The ll_destroyXXX() functions are checking if their parameter is
NULL, so there is no need for the caller to check it.

git-svn-id: http://www.streamboard.tv/svn/oscam/trunk@10436 4b0bc96b-bc66-0410-9d44-ebda105a78c1