[dvbapi] Reworked "ca_pmt_cmd_id" skipping logic
According to the spec (https://www.dvb.org/resources/public/standards/En50221.V1.pdf), the "ca_pmt_cmd_id"
value shall be present inside the ES info loop. Unfortunately, each dvbapi client does each own thing. The
only reliable way to detect when this info is present in the CA PMT is to actually check its value, like it
was done in the old code.

Most boxes always send the typical value of 0x01 (ok_descrambling) and the above check works fine because
there is no descriptor_tag 0x01 to get confused with. On the contrary, for enigma2, values like 0x03 (query),
0x04 (not_selected) are possible (not yet, but soon). Skipping these higher values the same way will not work,
because descriptors with such tags can exist, so a different check is implemented for enigma2.

In practice, this commit fixes the VDR dvbapi plugin, as well as the enigma2 boxes with wrong dvbapi boxtype
setting ("none" instead on the correct "dreambox" - the setting is irrelevant now, but will make a difference
in the future, so make sure you set it correctly).

git-svn-id: http://www.streamboard.tv/svn/oscam/trunk@11537 4b0bc96b-bc66-0410-9d44-ebda105a78c1